rename server-messages -> serverMessages and connect client->server-main->server->client
