+++ src/modules/server/serverMessages.js
@@ -5,5 +5,8 @@
 // messages for redirect incoming events from server to IPC-Main
 module.exports = {
   UPDATE_CONNECTION_STATUS: "update-connection-status",
-  UPDATE_SETS: "update-sets"
+  UPDATE_SETS: "update-sets",
+  // TODO: improve naming
+  STATE_REQUEST: "state-request",
+  STATE_RESPONSE: "state-response"
 };
